microlearning_content_processor/requirements.txt
# Core AI/ML Libraries - Essential for transcription and processing
openai-whisper==20231117  # Audio transcription
moviepy==1.0.3           # Video/audio processing

# Natural Language Processing - Required for topic analysis
spacy==3.7.2            # NLP pipeline
# Note: Also run: python -m spacy download en_core_web_sm

# Data Processing & Validation - Core data structures
pydantic==2.4.2         # Data validation and models
numpy>=1.21.0           # Numerical processing

# System Compatibility & Utilities
pathlib2>=2.3.0        # Path handling
typing-extensions>=4.0.0 # Type hints compatibility

# Media Processing Support
ffmpeg-python>=0.2.0   # Video/audio codec support

# PyTorch Dependencies (required by Whisper)
torch>=1.9.0           # Deep learning framework
torchaudio>=0.9.0      # Audio processing for Whisper

# Optional: Enhanced NLP (uncomment if needed)
# nltk==3.8.1
# scikit-learn==1.3.0  
# transformers==4.34.0
# pandas==2.0.3

# Optional: CUDA support for GPU acceleration (uncomment if using GPU)
# torch>=1.9.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchaudio>=0.9.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Development/Testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
